msg/dpdk: link common_async_dpdk to shared lib
The common_async_dpdk library depends on the dpdk rte_eal.a file.
If the static library is used, the dependency is transferred to
libceph-common. By default, libceph-common uses the PUBLIC keyword,
the keyword PUBLIC determines that both the application and
libceph-common link to dpdk rte_eal.a. As a result, the global variable
in dpdk rte_eal.a has two copies. Change common_async_dpdk to a shared
library and use PRIVATE to link to dpdk rte_eal.a. The keyword PRIVATE
determines that only common_async_dpdk link to dpdk rte_eal.a.

Fixes: https://tracker.ceph.com/issues/42861

Signed-off-by: Chunsong Feng <fengchunsong@huawei.com>
Signed-off-by: luorixin <luorixin@huawei.com>